use bottom navigation by navigation component
