+++ app/src/main/java/com/example/jokerfinder/features/MainActivity.kt
@@ -1,0 +1,34 @@
+package com.example.jokerfinder.features
+
+import android.annotation.SuppressLint
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import androidx.navigation.NavController
+import androidx.navigation.Navigation
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.NavigationUI
+import com.example.jokerfinder.R
+import com.example.jokerfinder.base.di.BaseActivity
+import com.google.android.material.bottomnavigation.BottomNavigationView
+
+
+class MainActivity : BaseActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        val navController = Navigation.findNavController(this, R.id.nav_host_fragment)
+        setUpBottomNav(navController)
+    }
+
+    private fun setUpBottomNav(navController: NavController) {
+        val bottomNavigationView = findViewById<BottomNavigationView>(R.id.bottom_navigation)
+        NavigationUI.setupWithNavController(bottomNavigationView,navController)
+    }
+}
+
+
+
+
+
