+++ app/src/main/java/com/example/jokerfinder/features/about/AboutUsFragment.kt
@@ -1,0 +1,25 @@
+package com.example.jokerfinder.features.about
+
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.jokerfinder.R
+
+/**
+ * A simple [Fragment] subclass.
+ */
+class AboutUsFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_about_us, container, false)
+    }
+
+
+}
