+++ app/src/main/res/menu/bottom_navigation.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/searchMovieFragment"
+        android:icon="@drawable/ic_home_black_24dp"
+        android:title="Home"
+        />
+
+    <item
+            android:id="@+id/favoriteMovieFragment"
+            android:icon="@drawable/ic_favorite_black_24dp"
+            android:title="Favorites"
+            />
+
+    <item
+            android:id="@+id/aboutUsFragment"
+            android:icon="@drawable/ic_info_black_24dp"
+            android:title="About us"
+            />
+
+
+</menu>
