+++ app/src/main/res/values/strings.xml
@@ -6,4 +6,7 @@
     <string name="character">Character :</string>
     <string name="released">Released :</string>
     <string name="vote_count">Vote Count :</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
